v0.9.0: namespace hygiene, build warnings cleanup, and documentation overhaul
Namespace migration:
- shell_*() → shell::init/handle_char/prompt/main()
- page2pa/page2kva/pa2page/kva2page/alloc_pages/free_pages → pmm::
- get_pte/page_insert/pgdir_alloc_page/tlb_invl → pmm::
- swapfs_init/read/write, find_vaddr_for_page → swap::
- InterruptsGuard → intr::Guard

Header refactoring:
- Split seg.h into seg.h (GDT-only) + new asm/memlayout.h (memory layout, E820)
- Remove obsolete e820.h/e820.cpp (inlined into pmm.cpp via boot_info)

Build system fixes:
- Eliminate all compiler warnings (6): type mismatches, unused functions
- Eliminate all linker warnings: .note.GNU-stack, --no-warn-rwx-segments
- Remove unused HOSTCC/HOSTCFLAGS, deduplicate boot CFLAGS

Cleanup:
- Extend .gitignore (*.o, *.d, *.bin, *.img, *.EFI, .vscode/, etc.)
- Delete junk files (-c, scripts/bin/)
- Fix stale .vscode/settings.json entries
- Fix README.md paths (tools/ → scripts/, make qemu-debug → debug-qemu)
- Update docs/TODO.md: correct paths, check completed items, update tree
- Resolve all DEVELOPMENT.md TODO items (high/medium/low)

+++ kernel/cons/stdio.cpp
@@ -42,7 +42,7 @@ void print_digit(uint64_t num, uint32_t base, int width, char padc) {
 
 void print_digit_no_pad(uint64_t num, int base) {
     uint32_t mod = 0;
-    if (num >= base) {
+    if (num >= static_cast<uint64_t>(base)) {
         mod = do_div(num, base);
         print_digit_no_pad(num, base);
     } else {
